+++ BridgeCareApp/BridgeCare/Controllers/HistoricalAttributesByYearController.cs
@@ -10,9 +10,9 @@ namespace BridgeCare.Controllers
     public class HistoricalAttributesByYearController : ApiController
     {
         private readonly BridgeCareContext db;
-        private readonly IAttributeByYear attributes;
+        private readonly IAttributesByYear attributes;
 
-        public HistoricalAttributesByYearController(IAttributeByYear attributeByYear, BridgeCareContext context)
+        public HistoricalAttributesByYearController(IAttributesByYear attributeByYear, BridgeCareContext context)
         {
             attributes = attributeByYear ?? throw new ArgumentNullException(nameof(attributeByYear));
             db = context ?? throw new ArgumentNullException(nameof(context));
